MGMT-3456 - Will not attempt to format bootable disks that look like installation media (#862)
* MGMT-3456 - Added is_installation_media field to disks

* MGMT-3456 - Will not attempt to format bootable disks that look like installation media

+++ internal/host/installcmd.go
@@ -213,7 +213,7 @@ func (i *installCmd) getDiskUnbootableCmd(ctx context.Context, host models.Host)
 	formatCmds := ""
 	for _, disk := range inventory.Disks {
 		isFcIscsi := strings.Contains(disk.ByPath, "-fc-") || strings.Contains(disk.ByPath, "-iscsi-")
-		if disk.Bootable && !isFcIscsi {
+		if disk.Bootable && !isFcIscsi && !disk.IsInstallationMedia {
 			dev := GetDeviceFullName(disk.Name)
 			formatCmds += fmt.Sprintf("dd if=/dev/zero of=%s bs=512 count=1 ; ", dev)
 			i.eventsHandler.AddEvent(
